virtual_attendance_system-main/requirements.txt
cmake
face-recognition
numpy==1.23.1
opencv-python
Flask==2.0.1
flask-restx==1.0.3
tensorflow
pandas
matplotlib
scikit-learn
keras
keras_applications
keras-vggface
Flask-Bcrypt
Werkzeug==2.0.1
streamlit-webcam-example
streamlit-webrtc
streamlit
Flask-Script==2.0.3
Flask-Migrate==2.6.0
Flask-Session